Remove workspace name from vmimport role
osbuild-composer does not specify the vmimport role when it talks to the
AWS API, and this means that the API defaults to an IAM role called
`vmimport`. Until we make that configurable, we need to ensure this role
is always called `vmimport`.

This needs to be reverted once osbuild/osbuild-composer#1170 has a
solution.

Signed-off-by: Major Hayden <major@redhat.com>