Provision GCP credentials on external workers
The latest version of osbuild-composer allows to import built images into
GCP. This functionality is intended to be used by the Image Builder
service. In order for the worker to do necessary GCP API calls, it needs to
authenticate. Since we are still deploying workers only in AWS, add the
necessary GCP Service Account credentials on external workers. This won't be
necessary, once we start provisioning workers also in GCP with proper
roles assigned to the worker's instance.

Signed-off-by: Tomas Hozza <thozza@redhat.com>