rpmrepo: implement API Gateway REST API
Add the API-Gateway and AWS Lambda integration of the RPMrepo REST API.
This expects the python-sources of the gateway to be provided as zip
file in the dedicated S3 bucket. The sources are taken from:

    https://github.com/osbuild/rpmrepo

The API-Gateway provides a simple {proxy+} hook and forwards all
requests to the AWS Lambda. This way, we basically shortcut most of the
API Gateway features and just implement all handlers ourselves (which
also follows most recommendations you find on the web).

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>