rpmrepo: fix SSL certificates by avoiding dots
AWS uses `*.s3.amazonaws.com` as altName in their SSL certificates. This
means only a single sub-label is covered by the cert (recursive or
nested matching is not supported by the underlying RFC). If we now use
bucket-names with dots, they will map to multi-label domain names and
thus cannot be accessed via SSL.

Fix this by moving the RPMrepo bucket. We have not deployed this bucket,
yet, so it can safely be moved.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>