rpmrepo: disentangle external input
Rather than storing external dependencies in buckets we create inside
terraform, we now disentangle the two and require an extra bucket to be
maintained outside of terraform.

The problem is, terraform currently creates our rpmrepo-storage bucket,
but further terraform resources depend on code-tarballs to be stored in
that bucket (e.g., AWS-Lambda requires code to be stored in an S3
bucket). This creates a circular dependency, which breaks new
deployments.

Avoid this and require an `rpmrepo-external` bucket to be separately
maintained, with access granted to our account. Then pull in the sources
from there.

While at it, turn those external requirements into variables so they are
not configured somewhere deep down as magic values in the code.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>