rpmrepo: add batch compute environment
Create an AWS Batch compute environment, job-queue, and IAM policies to
queue snapshot operations with the AWS Batch technology.

This still lacks the job-definition for the snapshot jobs. Those still
have to be created manually via the management console. Once they are
stable, we can add them here as well.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>